% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_anime_data.R
\docType{data}
\name{anime_data}
\alias{anime_data}
\title{Sample Anime Dataset}
\format{
A data frame with 20 rows and 5 columns:
\describe{
\item{Title}{Anime title}
\item{Genre}{Genre of the anime}
\item{Studio}{Production studio}
\item{Rating}{Average user rating (out of 10)}
\item{Year}{Year the anime was released}
}
}
\source{
Created for the AnimeFunPkg R package
}
\usage{
anime_data
}
\description{
A sample dataset containing anime titles, genres, studios, ratings, and release years.
}
\keyword{datasets}
